Updating my fork from runelite's original (#1)
* Add missing toggle for clan chat icons

Signed-off-by: Tomas Slusny <slusnucky@gmail.com>

* Clanchat: Show amount of members near you in infobox

* Fix RAIDS_POINTS_INFOBOX widget ID

Signed-off-by: Tomas Slusny <slusnucky@gmail.com>

* Make raids widget moveable

Signed-off-by: Tomas Slusny <slusnucky@gmail.com>

* Remove raids points overlay

As we can just make raids overlay moveable there is no need to have our
own.

Signed-off-by: Tomas Slusny <slusnucky@gmail.com>

* Null-check remote configuration in config manager

This field can be null if the config service do not find absolutely
anything in config table for the user.

Signed-off-by: Tomas Slusny <slusnucky@gmail.com>

* Change logged-in redir to redirect to HTTPS

Signed-off-by: Tomas Slusny <slusnucky@gmail.com>

* idle notifier: Add support for grinding with pestle and mortar

* worldpoint: Fix plane value in toLocalInstance

Fixes runelite/runelite#7985
Fixes runelite/runelite#8045

* daily task indicator: Fix plugin not checking tasks when first enabled

* Add Hydra shortcut pipes to agility plugin (#8052)

Fixes #8041

* Move session events from api to client

* api: add changed varp index to VarbitChanged event

* ground markers plugin: Clean up legacy code

This includes various code quality improvements such as stronger access
control, enforcing newlines at EOF, naming magic numbers and strings,
and utilizing higher efficiency methods and class constructors.

* worldpoint: Add static fromRegion method

* ground markers: Fix startup and shutdown bug

When logging in to the game with the plugin disabled, enabling it will
not cause points to be filled, as the game state trigger will not be
reached. This commit properly loads points on plugin startup and clears
them on shutdown.

* worldpoint: Add getRegionX() and getRegionY()

* ground markers: Allow different colored markers

This adds a color indicator per marker, saved to config, allowing
multiple markers to be different colors. In addition, a configuration
option is added to switch between per-tile color display or current
configured color display. (that is, new or existing behavior)

This change will set a color in config for old markers with null colors
when those markers are loaded.

Fixes runelite/runelite#3395

* Clanchat: added onClanChanged event to count players already in scene

* ground markers: Fix marking non-marked tiles

This fixes the bug introduced in runelite/runelite#5890 where attempting
to mark a non-marked tile while having the "Remember color per tile"
config disabled would prevent new tiles from being marked. Instead, it
properly ensures that unmarking tiles with that config disabled will not
re-mark it with an updated color, causing a second unmark to be needed.

* api: Add EnumID class

* api: Add unlocked music tracks to VarPlayer

* api: Add getKeys() to EnumComposition

* client: Add Music Track Indicator plugin

* runescape-api: make RSNameableContainer a generic type

* Remove Monkey Madness I requirement from Kourend Diary.

Creating the tablet does not require quest completion.

* rogues den: Update plugin to use ItemContainerChanged event

* Update CheckStyle XML dtd link to one that exists (#8079)

* Update suppressions.xml dtd link to one that exists

Signed-off-by: Tomas Slusny <slusnucky@gmail.com>

* api: add method to remove MessageNodes

Co-authored-by: trimbe <benno.tremback@gmail.com>

* api: add clan member join and leave events

Co-authored-by: trimbe <benno.tremback@gmail.com>

* api: add methods to get clan owner and chat name

Co-authored-by: trimbe <benno.tremback@gmail.com>

* clanchat plugin: add join/leave messages

Co-authored-by: trimbe <benno.tremback@gmail.com>

* Replace assert keyword with JUnit assert functions

* mixins: use a ClanMember for ClanMemberJoined/Left

* clanchat: retrieve rank from ClanMember rather than ClanManager

Ranks would be cached at the time of join/leave which was incorrect. Using the rank from ClanMember will always be correct and won't require the user to have spoken recently to cache their rank.

* clanchat: respect rank icon config for join/leave messages

* clanchat plugin: add CC rank icons to oublic and private messages

* Clanchat: clear counter on login/connection lost instead of loading (#8068)

* Fix SpringBootWebApplication auto-configuration annotation

The exclude is supposed to be on SpringBootApplication annotation,
otherwise IntelliJ Spring plugin errors, and also SpringBootApplication
already contains EnableAutoConfiguration.

Signed-off-by: Tomas Slusny <slusnucky@gmail.com>

* Migrate SpringBootWebApplicationTest to spring profiles

- Instead of replacing configuration with dev configuration on classpath,
create dev profile and use that
- Move common properties to application.yaml

Signed-off-by: Tomas Slusny <slusnucky@gmail.com>

* Migrate Spring unit tests to spring profiles

Instead of replacing configuration with test configuration on classpath,
create test profile and use that.

Signed-off-by: Tomas Slusny <slusnucky@gmail.com>

* Enable checkstyle on test sources